+++ app/actions/websocket/event.test.ts
@@ -2,7 +2,7 @@
 // See LICENSE.txt for license information.
 
 import * as bookmark from '@actions/local/channel_bookmark';
-import * as scheduledPost from '@actions/local/scheduled_post';
+import * as scheduledPost from '@actions/websocket/scheduled_post';
 import * as calls from '@calls/connection/websocket_event_handlers';
 import {WebsocketEvents} from '@constants';
 
@@ -34,7 +34,7 @@ jest.mock('./threads');
 jest.mock('@calls/connection/websocket_event_handlers');
 jest.mock('./group');
 jest.mock('@actions/local/channel_bookmark');
-jest.mock('@actions/local/scheduled_post');
+jest.mock('@actions/websocket/scheduled_post');
 
 describe('handleWebSocketEvent', () => {
     const serverUrl = 'https://example.com';
